Add support for serde_json's arbitrary precision
